requirements/base.txt
boto3==1.24.61
celery==5.2.7
channels-redis==3.4.1
channels==3.0.5
django-cors-headers==3.13.0
django-graphql-jwt==0.3.4
django-redis==5.2.0
django-redis-cache==3.0.1
django-rq==2.5.1
Django==4.1.0
graphene==3.1
graphene-django==3.0.0b7
graphene-django-extras==0.5.2
graphene-file-upload==1.3.0
mailjet-rest==1.3.4
mysqlclient==2.1.1
pyOpenSSL==22.0.0
python-decouple==3.6
python-graphql-client==0.4.3
Pillow==9.2.0